files/requirements.txt
sqlalchemy>=1.4
asyncpg==0.27.0
aiosqlite>=0.17.0
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
pydantic>=1.10.7
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
alembic>=1.10.0
pytest>=7.0.0
httpx>=0.24.0